+++ src/builder/utils/extractTraits.ts
@@ -1,7 +1,7 @@
 import {BuildTimeConfig} from '../types';
 
-export function extractTraits<Result, Trait extends string, MappedResult, Parameters extends any[]>(
-    buildTimeConfig?: BuildTimeConfig<Result, Trait, MappedResult, Parameters>,
+export function extractTraits<Origin, Result = Origin, PostBuildResult = Result, Trait extends string = string>(
+    buildTimeConfig?: BuildTimeConfig<Origin, Result, PostBuildResult, Trait>,
 ) {
     const traits = buildTimeConfig?.traits;
     return Array.isArray(traits) ? traits : traits ? [traits] : [];