+++ src/index.ts
@@ -17,4 +17,8 @@ export type {
     FieldsConfigurationGeneratorFunction,
     TraitsConfiguration,
     Overrides,
+    FieldType,
 } from './builder/types';
+
+export {resetable} from 'reset/resetable';
+export type {ResetSignal} from 'reset/ResetSignal';